LayoutTests/fast/spatial-navigation/snav-imagemap-area-without-image-expected.txt




PASS gFocusedDocument.activeElement.getAttribute("id") is "3"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
This test tests that areas of an imagemap without an image using it, are not focusable, thus can not be reached with spatial navigation.
